+++ gdsfactory/tests/test_get_bundle_west_to_north.py
@@ -11,14 +11,10 @@ def test_get_bundle_west_to_north(
     lengths = {}
 
     c = gf.Component("test_get_bundle_west_to_north")
-    w = h = 10
+    pad = gf.partial(gf.c.pad, width=10, height=10)
     c = gf.Component()
-    pad_south = gf.components.pad_array(
-        port_names=["e4"], pitch=15.0, pad_settings=dict(width=w, height=h), n=3
-    )
-    pad_north = gf.components.pad_array(
-        port_names=["e2"], pitch=15.0, pad_settings=dict(width=w, height=h), n=3
-    )
+    pad_south = gf.components.pad_array(port_names=["e4"], pitch=15.0, n=3, pad=pad)
+    pad_north = gf.components.pad_array(port_names=["e2"], pitch=15.0, n=3, pad=pad)
     pl = c << pad_south
     pb = c << pad_north
     pl.rotate(90)